Fix: implement Epic F deterministic IDs and suite splits
