+++ tests/e2e/platform/f-2-canonical-comms-event-model-and-event-store.atdd.spec.ts
@@ -66,7 +66,7 @@ const buildThreadUrl = (
 };
 
 test.describe(
-  'Story f.2 Canonical Comms Event Model and Event Store (ATDD E2E RED)',
+  'Story f.2 Canonical Comms Event Model and Event Store (ATDD E2E Timeline)',
   () => {
     test(
       '[P0] thread detail timeline renders canonical event rows with provider-neutral labels and deterministic chronological ordering @P0',
@@ -178,188 +178,6 @@ test.describe(
         events.forEach((event) => {
           assertProviderSpecificLeakageRemoved(event.payload);
         });
-      },
-    );
-
-    test(
-      '[P1] operator debug events panel filters by aggregate id and canonical event type while preserving deterministic result ordering @P1',
-      async ({ page, request }) => {
-        const context = createStoryF2Context();
-        const operatorHeaders = createStoryF2Headers(context, {
-          role: 'ORGUNIT_MEMBER',
-          orgUnitMemberships: [context.orgUnitId],
-        });
-        const adminHeaders = createStoryF2Headers(context, {
-          role: 'ORGUNIT_ADMIN',
-          userId: context.adminUserId,
-          orgUnitMemberships: [context.orgUnitId],
-        });
-        await login(page);
-
-        await page.goto(
-          buildThreadUrl(context, {
-            threadId: context.threadIds.claimed,
-            actorUserId: context.userId,
-            tenantRole: 'ORGUNIT_MEMBER',
-            orgUnitMemberships: [context.orgUnitId],
-            providerKey: context.providers.enabledPrimary,
-          }),
-        );
-
-        await expect(page.getByTestId('connectshyft-thread-detail')).toBeVisible();
-        await expect(page.getByTestId('connectshyft-thread-state-chip')).toHaveText('CLAIMED');
-
-        await apiRequest(request, {
-          method: 'POST',
-          path: context.paths.inboundWebhook,
-          headers: adminHeaders,
-          data: {
-            eventType: 'voice.connected',
-            threadId: context.threadIds.claimed,
-            orgUnitId: context.orgUnitId,
-            tenantId: context.tenantId,
-            providerKey: context.providers.enabledPrimary,
-            providerEventId: `provider-event-f2-atdd-filter-${randomUUID().slice(0, 8)}`,
-          },
-        });
-
-        const filterQuery = new URLSearchParams({
-          tenantId: context.tenantId,
-          orgUnitId: context.orgUnitId,
-          aggregateId: context.threadIds.claimed,
-          aggregateType: 'Thread',
-          eventType: context.canonicalEventTypes.callConnected,
-          limit: '50',
-        });
-        const firstResponse = await apiRequest(request, {
-          method: 'GET',
-          path: `${context.paths.events}?${filterQuery.toString()}`,
-          headers: operatorHeaders,
-        });
-        const secondResponse = await apiRequest(request, {
-          method: 'GET',
-          path: `${context.paths.events}?${filterQuery.toString()}`,
-          headers: operatorHeaders,
-        });
-
-        expect(firstResponse.status()).toBe(200);
-        expect(secondResponse.status()).toBe(200);
-
-        const firstBody = await firstResponse.json();
-        const secondBody = await secondResponse.json();
-        expect(hasRequiredEnvelopeKeys(firstBody)).toBe(true);
-        expect(hasRequiredEnvelopeKeys(secondBody)).toBe(true);
-        expect(firstBody).toMatchObject({
-          ok: true,
-          code: 'CONNECTSHYFT_CANONICAL_EVENTS_LISTED',
-          data: {
-            deterministic: true,
-            providerNeutral: true,
-          },
-        });
-        expect(secondBody).toMatchObject({
-          ok: true,
-          code: 'CONNECTSHYFT_CANONICAL_EVENTS_LISTED',
-          data: {
-            deterministic: true,
-            providerNeutral: true,
-          },
-        });
-
-        const firstEvents = firstBody.data.events as CanonicalEventRecord[];
-        const secondEvents = secondBody.data.events as CanonicalEventRecord[];
-        expect(firstEvents.length).toBeGreaterThan(0);
-        expect(firstEvents).toEqual(sortCanonically(firstEvents));
-        expect(secondEvents).toEqual(sortCanonically(secondEvents));
-        expect(firstEvents).toEqual(secondEvents);
-        firstEvents.forEach((event) => {
-          assertProviderSpecificLeakageRemoved(event.payload);
-          expect(event.eventType).toBe(context.canonicalEventTypes.callConnected);
-        });
-      },
-    );
-
-    test(
-      '[P1] status and timeline summary chips remain provider-neutral and event-derived without twilio or telnyx naming leakage @P1',
-      async ({ page, request }) => {
-        const context = createStoryF2Context();
-        const operatorHeaders = createStoryF2Headers(context, {
-          role: 'ORGUNIT_MEMBER',
-          orgUnitMemberships: [context.orgUnitId],
-        });
-        const adminHeaders = createStoryF2Headers(context, {
-          role: 'ORGUNIT_ADMIN',
-          userId: context.adminUserId,
-          orgUnitMemberships: [context.orgUnitId],
-        });
-        await login(page);
-
-        await page.goto(
-          buildThreadUrl(context, {
-            threadId: context.threadIds.unclaimed,
-            actorUserId: context.userId,
-            tenantRole: 'ORGUNIT_MEMBER',
-            orgUnitMemberships: [context.orgUnitId],
-            providerKey: context.providers.enabledPrimary,
-          }),
-        );
-
-        await expect(page.getByTestId('connectshyft-thread-detail')).toBeVisible();
-
-        await apiRequest(request, {
-          method: 'POST',
-          path: `${context.paths.threads}/${context.threadIds.unclaimed}/call`,
-          headers: operatorHeaders,
-          data: {
-            orgUnitId: context.orgUnitId,
-            providerKey: context.providers.enabledPrimary,
-          },
-        });
-        await apiRequest(request, {
-          method: 'POST',
-          path: context.paths.inboundWebhook,
-          headers: adminHeaders,
-          data: {
-            eventType: 'voice.fallback',
-            threadId: context.threadIds.unclaimed,
-            orgUnitId: context.orgUnitId,
-            tenantId: context.tenantId,
-            providerKey: context.providers.enabledPrimary,
-            providerEventId: `provider-event-f2-atdd-summary-${randomUUID().slice(0, 8)}`,
-          },
-        });
-
-        const detailResponse = await apiRequest(request, {
-          method: 'GET',
-          path: `${context.paths.threads}/${context.threadIds.unclaimed}`,
-          headers: operatorHeaders,
-        });
-
-        expect(detailResponse.status()).toBe(200);
-        const detailBody = await detailResponse.json();
-        expect(hasRequiredEnvelopeKeys(detailBody)).toBe(true);
-        expect(detailBody).toMatchObject({
-          ok: true,
-          code: 'CONNECTSHYFT_THREAD_DETAIL_LOADED',
-          data: {
-            thread: {
-              threadId: context.threadIds.unclaimed,
-              providerNeutral: true,
-              statusDerivedFromCanonicalEvents: true,
-              timeline: expect.any(Array),
-            },
-          },
-        });
-
-        const timeline = detailBody.data.thread.timeline as CanonicalEventRecord[];
-        expect(timeline.length).toBeGreaterThan(0);
-        expect(timeline).toEqual(sortCanonically(timeline));
-        timeline.forEach((event) => {
-          assertProviderSpecificLeakageRemoved(event.payload);
-        });
-
-        await expect(page.getByText(/twilio/i)).toHaveCount(0);
-        await expect(page.getByText(/telnyx/i)).toHaveCount(0);
       },
     );
   },
